Move larp relation to base StoryObject
